Remove intermediate handoff pid
Instead of using the handoff pid (`HPid`) "trick" in order to detect
vnode death and thus complicating the sender code make the handoff
manager be responsible for monitoring the vnode and kill the sender if
vnode death is detected.

As a refresher, the reason the `HPid` was introduced in the first
place is because vnode master `sync_command` will wait infinitely if
the vnode dies.  See commit `c44b086fd924f72809913c8059f4f02596b41548`
for more info.