Check `Reason` in vnode terminate
If the worker pool is busy doing something, or `Mod:terminate` does
blocking work, it could keep the vnode process in the `terminate`
callback for up to 60s.  This is bad because for that 60s msgs will be
queued up on the mailbox but then at the end of terminate they will
just drop.  In the case of normal termination you probably want to let
the worker finish it's work but a non-normal exit should bring down
the vnode quickly to allow it to restart.